refactor: feed 테이블의 like_count 타입을 BIGINT로 변경

+++ src/main/resources/schema-mysql.sql
@@ -115,7 +115,7 @@ CREATE TABLE feed (
     image_url   VARCHAR(500)        NOT NULL, -- 일상: 일상사진, 정보공유: 대표사진
     feed_type   ENUM('INFO', 'DAILY') NOT NULL,
     category_id BIGINT UNSIGNED     NULL,
-    like_count  INT     UNSIGNED    NOT NULL DEFAULT 0,
+    like_count  BIGINT     UNSIGNED    NOT NULL DEFAULT 0,
     created_at  TIMESTAMP           NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at  TIMESTAMP       NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     PRIMARY KEY (id),
